Added api requests analytics page
