+++ studio/components/ui/Charts/Charts.constants.ts
@@ -17,3 +17,17 @@ export const USAGE_COLORS = {
   404: 'var(--colors-amber7)',
   500: 'var(--colors-red9)',
 }
+
+
+
+export const CHART_INTERVALS = [
+  {
+    key: 'minutely',
+    label: '60 minutes',
+    startValue: 1,
+    startUnit: 'hour',
+    format: 'MMM D, h:mma',
+  },
+  { key: 'hourly', label: '24 hours', startValue: 24, startUnit: 'hour', format: 'MMM D, ha' },
+  { key: 'daily', label: '7 days', startValue: 7, startUnit: 'day', format: 'MMM D' },
+]
